refactor(ruby): unify test structure with shared ValkeyTests modules
- Rename test/standalone/ to test/valkey/ for shared modules
- Convert test classes to ValkeyTests::* modules (like Lint::*)
- Create new test/standalone/ with test classes that include modules
- Update test/cluster/cluster_commands_test.rb to include ValkeyTests modules
- Update test_helper.rb to load valkey shared test modules
- Fix assert_not_nil -> refute_nil for Minitest compatibility
