--- test/standalone/connection_commands_test.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-require "test_helper"
-
-class TestConnectionCommands < Minitest::Test
-  include Helper::Client
-  include Lint::ConnectionCommands
-end
