--- test/standalone/connection_options_test.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-require "test_helper"
-
-class TestConnectionOptions < Minitest::Test
-  include Helper::Client
-  include Lint::ConnectionOptions
-end
