--- test/standalone/hyper_log_log_commands_test.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-require "test_helper"
-
-class TestHyperLogLogCommands < Minitest::Test
-  include Helper::Client
-  include Lint::HyperLogLog
-end
