--- test/standalone/list_commands_test.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-require "test_helper"
-
-class TestListCommands < Minitest::Test
-  include Helper::Client
-  include Lint::Lists
-end
