--- test/standalone/module_commands_test.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-require "test_helper"
-
-class TestModuleCommands < Minitest::Test
-  include Helper::Client
-  include Lint::ModuleCommands
-end
