--- test/standalone/pubsub_commands_test.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-require_relative "../test_helper"
-
-class TestPubSubCommands < Minitest::Test
-  include Helper::Client
-  include Lint::PubSubCommands
-end
