--- test/standalone/sorted_set_commands_test.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-require "test_helper"
-
-class SortedSetCommandsTest < Minitest::Test
-  include Helper::Client
-  include Lint::SortedSetCommands
-end
