--- test/standalone/string_commands_test.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-require "test_helper"
-
-class TestStringCommands < Minitest::Test
-  include Helper::Client
-  include Lint::StringCommands
-end
