--- test/standalone/transaction_commands_test.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-require "test_helper"
-
-class TestTransactionCommands < Minitest::Test
-  include Helper::Client
-  include Lint::TransactionCommands
-end
