--- test/standalone/vector_search_commands_test.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-require "test_helper"
-
-class TestVectorSearchCommands < Minitest::Test
-  include Helper::Client
-  include Lint::VectorSearchCommands
-end
